requirements.txt
pandas
requests
numpy
streamlit
scikit_learn
tensorflow
python-dotenv
plotly